Store API - Removing Content-Type, Accept headers and adding security definitions

+++ components/apimgt/org.wso2.carbon.apimgt.rest.api.store/src/main/java/org/wso2/carbon/apimgt/rest/api/store/mappings/TierMappingUtil.java
@@ -70,13 +70,13 @@ public class TierMappingUtil {
         dto.setName(tier.getPolicyName());
         dto.setDescription(tier.getDescription());
         dto.setTierLevel(TierDTO.TierLevelEnum.valueOf(StringUtils.upperCase(tierLevel)));
-        dto.setUnitTime(tier.getDefaultQuotaPolicy().getLimit().getUnitTime());
+        dto.setUnitTime((long) tier.getDefaultQuotaPolicy().getLimit().getUnitTime());
 
         Limit limit = tier.getDefaultQuotaPolicy().getLimit();
         if (limit instanceof RequestCountLimit) {
-            dto.setRequestCount(((RequestCountLimit) limit).getRequestCount());
+            dto.setRequestCount((long) (((RequestCountLimit) limit).getRequestCount()));
         } else if (limit instanceof BandwidthLimit) {
-            dto.setRequestCount(((BandwidthLimit) limit).getDataAmount());
+            dto.setRequestCount((long) (((BandwidthLimit) limit).getDataAmount()));
         }
         //// TODO: 08/12/16 More fields to map 
         return dto;
